requirements.txt
atomicwrites==1.4.1
attrs==22.2.0
black==22.8.0
bleach==4.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.5
cryptography==40.0.2
dataclasses==0.8
docutils==0.18.1
dynaconf==3.1.5
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
iniconfig==1.1.1
keyring==23.4.1
mypy-extensions==1.0.0
packaging==21.3
pathspec==0.9.0
pdfminer.six==20221105
pkginfo==1.9.6
platformdirs==2.4.0
pluggy==1.0.0
py==1.11.0
pycparser==2.21
Pygments==2.14.0
pyparsing==3.0.7
pytest==7.0.1
pywin32-ctypes==0.2.2
readme-renderer==34.0
requests==2.27.1
requests-toolbelt==1.0.0
rfc3986==1.5.0
six==1.16.0
tomli==1.2.3
tqdm==4.64.1
twine==3.8.0
typed-ast==1.5.5
typing_extensions==4.1.1
urllib3==1.26.16
webencodings==0.5.1
zipp==3.6.0
